pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-hallucination-detector-gym"
version = "1.0.0"
description = "An OpenEnv environment for training AI agents to detect and classify hallucinations in LLM-generated text"
readme = "README.md"
license = {text = "BSD-3-Clause"}
requires-python = ">=3.10"
authors = [
    {name = "Hallucination Detector Gym Team"},
]
keywords = ["openenv", "reinforcement-learning", "hallucination-detection", "ai-safety", "llm"]

dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.3",
    # Environment-specific dependencies
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "structlog>=23.0.0",
    "websocket-client>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.24.0",
]

[project.scripts]
# Server entry point — enables running via: uv run --project . server
server = "server.app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["hallucination_detector_gym*", "server*"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = ["tests"]
